Refactor ConnectorLike
