+++ japgolly/src/main/scala/react/japgolly/package.scala
@@ -1,9 +1,12 @@
 package eldis.redux.react
 
-import _root_.japgolly.scalajs.react._
 import org.scalajs.dom.raw.Element
+import scala.scalajs.js
+
+import _root_.japgolly.scalajs.react._
 
 import eldis.redux
+import base.JsWrapper
 
 /**
  * A wrapper over react-redux, compatible with japgolly's scalajs-react
@@ -17,15 +20,33 @@ package object japgolly {
    * See [[https://github.com/reactjs/react-redux/blob/master/docs/api.md#provider-store the react-redux documentation]]
    * for detailed description.
    */
-  val Provider = JapgollyImpl.Provider
+  object Provider {
 
-  type ProviderProps = JapgollyImpl.ProviderProps
-  val ProviderProps = JapgollyImpl.ProviderProps
+    def apply[S, A](store: redux.Store[S, A])(child: ReactNode) = {
+      val provider = base.Provider.asInstanceOf[JsComponentType[ProviderProps, js.Any, Element]]
+      React.createFactory(provider)(
+        ProviderProps(store), child
+      )
+    }
+  }
 
-  /** The function that maps the state and the dispatcher function to the component's properties */
-  type Connector[S, A, P, OP] = base.Connector[S, A, P, OP]
+  /** The properties of Provider component. */
+  @js.native
+  trait ProviderProps extends js.Object {
+    val store: js.Any = js.native
+  }
 
-  type ConnectedComponentFactory[Props, State, +Backend, +Node <: TopNode] = JapgollyImpl.ConnectedComponentFactory[Props, State, Backend, Node]
+  object ProviderProps {
+    def apply(store: js.Any): ProviderProps =
+      js.Dynamic.literal(
+        store = store
+      ).asInstanceOf[ProviderProps]
+  }
+
+  /** The factory that produces component's connected to the store. */
+  trait ConnectedComponentFactory[Props, State, +Backend, +Node <: TopNode] {
+    def apply(props: Props, children: ReactNode*): ReactComponentU[Props, State, Backend, Node]
+  }
 
   /**
    * Creates the connected to state component factory.
@@ -33,8 +54,19 @@ package object japgolly {
    * @param connector  The function that maps state and dispatcher function to component's properties
    * @param cls        The component's class
    */
-  @inline def connect[S, A, P, OP, S1, B](connector: redux.Dispatcher[A] => (S, OP) => P, cls: ReactClass[P, S1, B, Element]): ConnectedComponentFactory[OP, S1, B, Element] =
-    JapgollyImpl.connect(connector, cls)
+  @inline def connect[C, P, OP, S1, B](
+    connector: C, cls: ReactClass[P, S1, B, Element]
+  )(implicit C: base.ConnectorLike[C, P, OP]): ConnectedComponentFactory[OP, S1, B, Element] =
+
+    new ConnectedComponentFactory[OP, S1, B, Element] {
+      def apply(props: OP, children: ReactNode*) = {
+        React.createFactory(
+          base.connect[C.State, C.Action, Any, P, OP, ReactClass[?, S1, B, Element], WrapObj, WrapObj[P], WrapObj[OP]](
+            C(connector)
+          )(cls)
+        )(WrapObj(props), children)
+      }
+    }
 
   /**
    * Connects stateless component to the global state
@@ -44,9 +76,9 @@ package object japgolly {
    */
   @inline def connect[C, P, OP](connector: C, comp: FunctionalComponent[P])(
     implicit
-    C: base.ConnectorLike[C, _, _, P, OP]
+    C: base.ConnectorLike[C, P, OP]
   ): FunctionalComponent[OP] =
-    JapgollyImpl.connect(C(connector), comp)
+    base.connect(C(connector))(comp)
 
   /**
    * Connects stateless component with children to the global state
@@ -56,7 +88,13 @@ package object japgolly {
    */
   @inline def connect[C, P, OP](connector: C, comp: FunctionalComponent.WithChildren[P])(
     implicit
-    C: base.ConnectorLike[C, _, _, P, OP]
+    C: base.ConnectorLike[C, P, OP]
   ): FunctionalComponent.WithChildren[OP] =
-    JapgollyImpl.connect(C(connector), comp)
+    base.connect(C(connector))(comp)
+
+  private implicit val wrapperInstance =
+    new JsWrapper[WrapObj] {
+      override def wrap[A](a: A) = WrapObj(a)
+      override def unwrap[A](fa: WrapObj[A]) = fa.v
+    }
 }
